+++ Cargo.toml
@@ -1,16 +1,22 @@
 [workspace]
-    
-members = [ "crates/bcr-web-api", "crates/wildcat"]
+
+
+members = [ "crates/bcr-web-api", "crates/bcr-wdc-keys", "crates/wildcat"]
+
 
 [workspace.dependencies]
+anyhow = {version = "1.0"}
 async-trait = {version = "0.1"}
 axum = {version = "0.6", features = ["macros"]}
 bitcoin = {version = "0.32"}
 cdk = {version = "0.6"}
 chrono = {version = "0.4", features = ["serde"]}
 log = {version = "0.4", features = ["serde"]}
+mockall = {version = "0.13"}
+rand = {version = "0.8"}
 rust_decimal = {version = "1.36"}
 serde = {version = "1.0", features = ["derive"]}
 surrealdb = {version = "2.2", features = ["kv-mem"]}
+thiserror = {version = "2.0"}
 tokio = {version = "1.4", features = ["macros", "rt-multi-thread"]}
 uuid = {version = "1.11", features = ["serde", "v4"]}
